Rename duplicate modules in GenerateHarness
If modules exist in both GenerateHarness and GenerateTop then the
modules in GenerateHarness get _inTestHarness appended to them until
they're unique.  Black boxes are not renamed, and neither is the top
module (which is eventually converted to a black box).

There's a few grep-based tests, but they might be kind of flaky.